updated the way affiliations get deleted through the api
