+++ api/endpoint-tests/endpoint.sh
@@ -2,6 +2,11 @@
 
 rm ./endpoint-tests/endpoint-data.json
 
+export NODE_ENV=test
+export API_URL=http://localhost:8081
+export CYPRESS_TESTS=false
+unset DEV_DB_NAME
+
 echo "[]" > endpoint-data.json
 
 ENDPOINT_COVERAGE_CAPTURE=true docker-compose -f docker-compose.endpoint-tests.yml up -d
