+++ api/routes/affiliations/index.js
@@ -1,19 +1,12 @@
 const get = require('./get');
 const post = require('./post');
 const patch = require('./patch');
-const del = require('./delete');
 
 module.exports = (
   app,
-  {
-    getEndpoint = get,
-    postEndpoint = post,
-    patchEndpoint = patch,
-    deleteEndpoint = del
-  } = {}
+  { getEndpoint = get, postEndpoint = post, patchEndpoint = patch } = {}
 ) => {
   getEndpoint(app);
   postEndpoint(app);
   patchEndpoint(app);
-  deleteEndpoint(app);
 };
