+++ api/routes/cypress/index.js
@@ -1,0 +1,5 @@
+const del = require('./delete');
+
+module.exports = (app, { deleteEndpoint = del } = {}) => {
+  deleteEndpoint(app);
+};
