+++ web/package.json
@@ -18,8 +18,8 @@
     "build-storybook": "npm run prebuild:storybook && build-storybook -o ../docs-build -c .storybook -s ./dist",
     "storybook": "npm run scaffold && start-storybook -s ./dist -p 6006",
     "preinstall": "npx npm-force-resolutions",
-    "cy:open": "./cypress/cypress-setup.sh && cypress open",
-    "cy:run": "./cypress/cypress-setup.sh && cypress run",
+    "cy:open": "./cypress/cypress-setup.sh open",
+    "cy:run": "./cypress/cypress-setup.sh run",
     "cy:run:ci": "cypress run"
   },
   "repository": {
